Implemented CreateSSH action and refactor
